redirect from index.ts instead so that we can import the ts moduleManifest
